86cynj382: Refactor TaskDTO object into dedicated model folder
